+++ src/components/Paper/index.ts
@@ -1,3 +1,3 @@
-// Paper from copy-ui @ 2025-10-17.
+// Paper from copy-ui @ 2025-10-22.
 
 export { Paper } from "./Paper";