+++ src/components/Paper/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-10-17: Initial version.
+- 2025-10-22: Add `withPadding` prop to control internal padding.